Add newlines in quotes and RTs for clarity
This changes the format for quotes and retweets when crossposted to
Mastodon, adding an empty line between the twitter username of the
author of the original tweet and the content. And in the case of quotes,
adding an empty line between the text of the quote and the name of the
original tweet author.
Fixes #147